changed return value of getNDF() in GFTrack.h to double to match the return value of getNDF() in GFAbsTrackRep.h

+++ trunk/core/GFTrack.h
@@ -237,7 +237,7 @@ class GFTrack : public TObject {
    *
    * Cardinal representation is used.
    */
-  unsigned int getNDF() const {return getCardinalRep()->getNDF();}
+  double getNDF() const {return getCardinalRep()->getNDF();}
 
   /** @brief Get chi2/NDF
    *
